+++ NEWS.md
@@ -1,4 +1,6 @@
 
+# renv (development version)
+
 # renv 1.2.2
 
 * Fixed an issue where `renv::install()` failed with named remotes in